move make_organization_and_user_with_token fixture to global scope
