+++ engine/conftest.py
@@ -765,3 +765,13 @@ def make_organization_and_region(make_organization, make_region):
         return organization, region
 
     return _make_organization_and_region
+
+
+@pytest.fixture()
+def make_organization_and_user_with_token(make_organization_and_user, make_public_api_token):
+    def _make_organization_and_user_with_token():
+        organization, user = make_organization_and_user()
+        _, token = make_public_api_token(user, organization)
+        return organization, user, token
+
+    return _make_organization_and_user_with_token
